[feat]: #77 - 페이지 초기 로드 때만 리마인드 인사이트 호출하도록 전역 플래그 설정
