+++ src/components/home/body/ChromeCalendar.tsx
@@ -4,12 +4,9 @@ import RenderCalendarBoard from './RenderCalendarBoard';
 import Left from '@svg/prev-icon.svg';
 import Right from '@svg/next-icon.svg';
 import Down from '@svg/down-icon.svg';
-import { useState } from 'react';
+import { useEffect, useState } from 'react';
 import { usePostReminderCalendar } from '@/hooks/api/useReminder';
-import { CalendarPostRequest, CalendarPostResponse } from '@/types/reminder';
-import { useCalendarPostResponseStore } from '@/store/reminder';
-import { Insight } from '@/types/insight';
-
+import { useFetchStore } from '@/store/reminder';
 interface Props {
   onClickModal: () => void;
   selectedDate: string;
@@ -21,28 +18,15 @@ const Calendar2 = ({ onClickModal, selectedDate, setSelectedDate }: Props) => {
   const splited = selectedDate.split('/');
   const [direction, setDirection] = useState<string>('');
   const { mutate } = usePostReminderCalendar();
-  const { setRecommendPostResponse } = useCalendarPostResponseStore();
+  const { hasFetched } = useFetchStore();
+
+  useEffect(() => {
+    !hasFetched && mutate(today);
+  }, []);
 
   const handleSelectDate = (date: string | null) => {
     date ? setSelectedDate(date) : setSelectedDate(today);
-    const postReminderCalendarRequest: CalendarPostRequest = {
-      requestDate: dayjs(date).format('YYYY-MM-DD'),
-    };
-    mutate(postReminderCalendarRequest, {
-      onSuccess: (data) => {
-        const dataWithFlag: CalendarPostResponse = {
-          ...data,
-          remindInsightList: data.remindInsightList.map((insight: Insight) => ({
-            ...insight,
-            isRead: false,
-          })),
-        };
-        setRecommendPostResponse(dataWithFlag);
-      },
-      onError: (error) => {
-        console.error(error);
-      },
-    });
+    mutate(String(date));
   };
 
   const handlePrevWeek = () => {
