+++ src/hooks/api/useReminder.ts
@@ -3,8 +3,11 @@ import {
   postReminderAnswer,
   postReminderCalendar,
 } from '@/api/reminder';
-import { AnswerPostRequest, CalendarPostRequest } from '@/types/reminder';
+import { Insight } from '@/types/insight';
+import { AnswerPostRequest, CalendarPostResponse } from '@/types/reminder';
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
+import { useCalendarPostResponseStore, useFetchStore } from '@/store/reminder';
+import dayjs from 'dayjs';
 
 export function useGetReminderQuestion() {
   return useQuery({
@@ -25,8 +28,24 @@ export function usePostReminderAnswer() {
 }
 
 export function usePostReminderCalendar() {
+  const { setRecommendPostResponse } = useCalendarPostResponseStore();
+  const { setHasFetched } = useFetchStore();
   return useMutation({
-    mutationFn: (calendarData: CalendarPostRequest) =>
-      postReminderCalendar(calendarData),
+    mutationFn: (date: string) =>
+      postReminderCalendar({ requestDate: dayjs(date).format('YYYY-MM-DD') }),
+    onSuccess: (data) => {
+      const dataWithFlag: CalendarPostResponse = {
+        ...data,
+        remindInsightList: data.remindInsightList.map((insight: Insight) => ({
+          ...insight,
+          isRead: false,
+        })),
+      };
+      setRecommendPostResponse(dataWithFlag);
+      setHasFetched(true);
+    },
+    onError: (error) => {
+      console.error(error);
+    },
   });
 }
